Change accelerators for section navigation
The accelerator keys for navigating the sections when not displaying
the whole source code conflicted with the well-known Ctrl+Right and
Ctrl+Left shortcuts for jumping to the next or previous word. Changed
to period and comma (i.e. same keys as < and >) as per the OS X front
end. Thanks to Jacqueline Ashwell for the suggestion.