Restart GConf daemon after installing schemas
The GConf daemon is only refreshed for the user that installs the
schemas, so when you do "sudo make install" to install the program
as root, then the schemas are not available to the regular user.
Sending the HUP signal to the GConf daemon refreshes it for all users.

Fixes Mantis #600.