Put hanging indent on tab-indented lines as well
The hanging indent needs to adapt to the number of tab characters at the
beginning of each line. This is done using GtkTextTags named indent-1,
indent-2, etc. that override the global hanging indent settings of the
text buffer. When inserting or deleting text, the relevant portion of
the text is re-scanned and these tags are re-applied.