Support GStreamer 0.10 or 1.0
Chimara now supports GStreamer 0.10 or 1.0; however, Inform is
constrained in which one it picks, since WebKit also uses GStreamer and
linking to both versions of GStreamer in the same process is not
allowed.

So, you can specify --with-sound=no or --with-sound=gstreamer to
configure, and it will examine which libraries WebKit links to in order
to determine which version of GStreamer to use.