Extensions pane
This adds the new Extensions pane, the flagship feature of this version
of Inform 7.

The three buttons in the toolbar are Home, to load the home page listing
installed extensions, Definitions, to load the index of definitions, and
Public Library, to load the Inform website's library of extensions.

The special links in the Public Library don't work yet; that is to be
solved in a future commit.