When choosing a panel, check which side is focused
This prevents a rare weirdness where if both sides are showing the same
pane, and a user action on the left side initiates a change to a
different view in the same pane (e.g. a different Index page), the view
would change on the right side instead. Now it changes on whatever side
had the focus.