Follow new indexing scheme for examples
Code examples are not contained in several HTML files anymore, but
instead all in the same one; they are revealed or hidden using
Javascript. So instead of indexing ex1234.html files and building one
DocText metadata object per HTML file, we now have to build several.
This makes the necessary changes to the indexing routine, but does not
yet update the UI, so double-clicking on an example in the search result
will not yet jump to that example.