Update UI of search window
The Recipe Book results are tinted yellow and the others white (and in
that case the rules hint doesn't work any more anyway, so remove it.)
The columns are reordered, and given sort strings; and the Example name,
if there is one, is prefixed to the search string context. The first two
columns now have a max width, and the text is ellipsized if it is
longer.