API for fetching installed extension version
This adds i7_app_get_extension_version() which returns whether an
installed extension is built-in to Inform and what version it identifies
itself as. Since this may be queried many times while running the Public
Library page's Javascript, the results should be cached; and it is easy
to do so in the tree store of extensions which is built.

To this end, extensions are not named based on their filename, but by
reading the first line of the extension file using is_valid_extension().
This also cuts out the need to remove .i7x from the file name.