Implement multiple downloads in Extensions pane
Clicking on the "Update All" or "Download All" buttons will download a
list of extensions using the downloadMultipleExtensions() function in
Javascript. The download progress will be shown in a dialog box. Any
single download that takes more than 15 seconds will be cancelled, but
the whole operation will continue. If the user presses the "Cancel"
button, on the other hand, the whole operation will be stopped. At the
end, the user is shown a dialog box with a list of errors if any
downloads failed.