Don't have the document unref itself
This was a bad idea. The owner of an object should unref it, not the
object itself when triggered by some action. This ended up causing a
segmentation fault because of some unref loop that I don't entirely
understand, but it's a bad idea so don't do it.

Fixes Mantis #1406.