Adapt GtkSourceView's auto-indent algorithm
The current algorithm was insufficient, as it auto-indented any text
that was inserted; even blocks of pasted text. Instead, adapt
GtkSourceView's auto-indent algorithm, which auto-indents on an explicit
key press of Enter or Return.

As a reminder, we don't use GtkSourceView's algorithm directly, because
it auto-indents on tabs and spaces, and we only want to auto-indent on
tabs. There's no way to change that behaviour in GtkSourceView.

Fixes Mantis #1497.
http://inform7.com/mantis/view.php?id=1497