Fix reverting to saved version
Since we now quit the app if all document windows are closed, add a
revert() method to I7Document that reloads the file from the latest
saved version instead of closing and opening the window.

This is also in line with the behaviour of GApplication which will quit
if all its application windows are closed, so that will make things
easier to port later.