Update inweb to 6M62 version
Inweb's executable was renamed from inwebc back to inweb, as it was
originally named a while back.

Inweb now requires a library web called inlib which lives in src/inlib.
Inlib doesn't need to be compiled separately, since it is compiled
directly into Inweb. However, we do add an Automake file to Inlib which
makes sure that Inlib is distributed along with the source code.