Replace \xNN escape sequences with \uNNNN
Instead of pre-coding our special characters into UTF-8, use \uNNNN
Unicode codepoint escape sequences. The \x sequences can go on without a
limit to the number of characters, so they can eat up subsequent text.
Previously we solved this with "\xEA""abc" but when we switch to Gettext
that is not parsed correctly.