Translate names of color schemes
GtkSourceView color scheme XML files use a now-deprecated convention
where the names of attributes and elements are prefixed with _ if they
are to be translated. I never understood that you were supposed to
translate them in code, but that does make sense because GtkSourceView
wouldn't be able to know the translation domain of an external scheme
that was installed.

Likewise, when displaying the names of GtkSourceView's built-in color
schemes, we have to translate them in the GtkSourceView translation
domain.

One can understand why that convention was deprecated in favor of
xml:lang attributes.