Use ITS Tool to translate MIME XML
According to the spec, MIME XML can use multiple <comment> elements with
xml:lang attributes, so that's what we generate with ITS Tool. Since we
don't use intltool for this anymore, we don't have to have <_comment>
elements.

ITS Tool should be optional for those building only from tarballs.
Therefore we distribute the files it outputs, and only put them in
MAINTAINERCLEANFILES.