Remove intltool
Intltool is not maintained, and everything it was needed for is now
handled by Gettext and ITS Tool. We can remove it.

Support for Glade files was added to Gettext 0.18.3, but on Debian
stable 0.19.3 is available, so we use that. (Gettext has an odd way of
doing things, in that if you request a certain version, it will use
exactly that version of itself, instead of the latest version. So we
want to use the latest possible version available to us.)