Better error message for missing ITS Tool
ITS Tool is only needed if compiling from a git clone, not from a
tarball. Previously a missing ITS Tool would produce a less-than-helpful
error message.

Now we change the dummy value from "notfound" to "itstool-notfound" so
it's more obvious what's going on if we hit the situation where the tool
is needed but missing; but also we try to detect the situation up front
in configure.

Thanks to Michael Baum for the suggestion.