Port all libraries to their GTK 3 equivalents
Replaces GTK 2 with GTK 3, GtkSourceView 2 with GtkSourceView 3, WebKit 1
with WebKit 2, GtkSpell 2 with GtkSpell 3, GooCanvas with GooCanvas 2, and
moves to newer commits of Chimara and Osxcart.

Makes all changes necessary to the code for this to compile and run,
although there are still warnings about deprecated functions, and not
everything works correctly.