Use G_DEFINE_TYPE_WITH_PRIVATE to define classes
The old way of doing it with g_type_class_add_private() is deprecated,
the new way compiles down to more efficient code anyway.

This does require some refactoring as it's less convenient to split
classes across multiple files, and that's probably a bad pattern anyway.

While we're at it, we make all the code consistently use "self" as the
name of the "self" pointer in all the methods.