Use GResources for data where possible
For everything that was previously fetched with i7_app_get_data_file(),
we should use GResource to compile it directly into the binary instead.
This requires a bit of refactoring.

The Documentation/ and Resources/ directories from the giant Inform data
files tarball should be included in this, so we need a bit of build glue
in order not to install them anymore, and instead generate a GResource
file for them. (They are too many to list by hand, and change from
release to release.)

GtkSourceView still requires the highlighting and color schemes to be in
actual directories on disk, and the mimetype files are meant to be
installed as well.