Re-save all Glade files
Newer Glade versions save things in a consistent format, so these should
no longer have arbitrary diffs every time they are saved.

Note that Glade does not save panel.ui correctly, so it's still best to
edit this one by hand for the time being. (It moves the webviews out of
order and adds a bunch of empty tabs to the end.)