transcript: Use height-for-width management for the renderer
gtk_cell_renderer_get_size() is deprecated, and we use the
height-for-width management functions to size the cell renderer.

I'm not sure if this was already the case in GTK 2, but the padding
properties are signed integers, so we make our custom padding properties
of the cell renderer signed as well.