Stop using gtk_menu_popup()
This function doesn't work on Wayland displays. Instead we use
gtk_menu_popup_at_pointer(), gtk_menu_popup_at_widget() and
gtk_menu_popup_at_rect() as appropriate. In order to use the latter on
skein nodes, we have to do a bit of refactoring to be able to get the
bounds of a node label instead of just its anchor point's coordinates.