app: Display Inform icon next to read-only extensions
In GTK 3 this is not shown by default, and is subject to a user
preference. Instead of using GtkImageMenuItem, create the image manually
and pack it into a box with the label, in order to ensure it's visible.
(It's not really an image menu item, more like a menu item with an
emblem.)