Remove use of GDK lock
We no longer try to call GUI functions from another thread while holding
the GDK lock. Instead, we use gdk_threads_add_idle() to queue it up
asynchronously. This is for the most part a straightforward refactor
except in story-compile.c where we have to do a bit of trickery to
analyze the I6 compiler output.