app: Pass app object to foreach_extension callbacks
Since the app object is used in some of these callbacks, we may as well
pass it in rather than fetching it with g_application_get_default(). This
allows these methods to be used before the application is run, which is
important for more loose coupling so that the code is more testable.