panel: Handle inform:///Extensions and inform:/Extensions
It seems Inform is not quite consistent about generating these. We have to
strip off the 'Extensions' path component and then look in a location with
or without it in the path.

This fixes going back to the main installed extensions page from the
generated extension documentation pages. The URL given is
inform:/Extensions/Extensions.html which must map to
~/Inform/Documentation/Extensions.html.