Rename app to "Inform7 IDE"
I named the app "GNOME Inform7" in 2006 not knowing any better. However,
GNOME is a trademark and cannot be used without permission. Naming an
application after the desktop environment that it runs on is not really a
best practice anyway. Therefore, renaming it to "Inform7 IDE" seems like a
good idea now.

Practically speaking, the branding inside the app has been "Inform" for a
long time already, so this is a mostly invisible change for users. It just
renames some filenames and some directories.

The repository and the Transifex project continue to be named
gnome-inform7 for the time being, although I plan to rename those as well.