panel: Use canvas scale to implement Skein font size
Instead of resetting the font size for each node when we change the
document font size in the preferences, use the 'scale' property of the
canvas to scale the whole thing.

This should be more performant, as well as just working better. Previously
the node background shapes didn't scale with the font size.