story: Delete extraneous empty folder in Save As/Save Copy
For I7 projects we have to use a file chooser dialog with the
CREATE_FOLDER action. When the user types a folder name into the file
chooser, GTK will create the folder automatically. If we then add
".inform" to the file name, and create _that_ folder, we'll be left with
an empty folder without ".inform". Delete this folder if it exists and is
empty. (The user could have typed an existing folder name.)

We don't need to check if the folder is a folder and not a regular file.
The file chooser will already check that. (This leads to the somewhat odd
situation that if you have a regular file "test" and you try to save a
project "test.inform" by typing "test", the file chooser will not let you
do that.)