transcript: Replace GtkTreeView with GtkListBox
Instead of using GtkTreeView and GtkCellRenderer to render the Transcript
pane, use the more modern GtkListBox widget. This allows the transcript
entries themselves to be widgets (I7TranscriptEntry), instead of cell
renderers.

Having I7TranscriptEntry be a widget also allows more closely following
the original app design of having "Bless", "Show knot", and "Play to here"
be buttons, instead of hiding away in a right-click popup menu.

I7Skein now implements the GListModel interface instead of GtkTreeModel.

The only complicated part is that scrolling to a particular row in
GtkListBox is not trivial. It is achieved by having the GtkListBoxRow grab
the focus, but that needs to be done asynchronously if the GtkListBox was
not shown in the containing GtkNotebook, or if the list model was being
changed at the same time.