Remove Page Setup menu item
The system print dialog allows embedding the page setup controls directly
in the print dialog. We don't need to have a separate menu item for this.

It appears that at one time you could use the page setup dialog to adjust
the margins on the paper, but that functionality already no longer exists
in the page setup dialog, so we're not losing anything by removing it.