panel: Style panel toolbar buttons
This adds the .inline-toolbar style class to the panel toolbars, so that
they stand out visually more than they used to. The buttons are given
borders, so it's clearer what they do.

We need to tweak the CSS a bit due to some assumptions made by GTK.
Apparently inline toolbars are usually placed directly under a view.