Remove AUTHORS.md and refresh authors list
The information in this file is duplicated in the app as well, and of
course in the Git log. It's better just to keep the information in the app
directly, I doubt anyone looks for this file specifically.

While we are doing this, update the authors list based on the latest
information from the Git log.