Change versioning scheme
We already renamed the package, so this seems like a good time to change
the versioning scheme as well. We'll use semantic versioning. Assuming
that I'd have named the previous versions 1.x, and that the port to GTK 3
and other non-obsolete GNOME libraries is a big enough change to warrant a
major version bump, that puts the upcoming version at 2.0.0.

We now make a distinction between the IDE version and the compiler version
and define a separate symbol for the latter.