Update Inform compiler
Integrates the new 10.1.0 version of the Inform compiler into the app.

Deletes the cBlorb, inform6, and inweb sources from the repo, as these are
now part of the Inform distribution.

Very few changes to the source are actually needed. Most of the changes
are to the build and packaging scripts.