story: Rewrite i7_story_open() to use autocleanup
There was a crash in i7_story_open() if the story failed to open, because
a "goto fail" was jumping across the initialization of an autoptr
variable. This does not work, because the initialization never happens but
the cleanup does.

This uses a neat trick that I learned from the Flatpak codebase, kind of
like implementing mozilla::ScopeExit in C-with-autocleanups, to clean up
the recent files entry automatically if a file failed to open, so that we
don't have to "goto fail" or remember to remove it at every exit point of
the function.