preferences: Remove extensions view
This view is redundant. There is a menu item that lets you install locally
downloaded extensions, or open an installed extension, and extensions can
also be installed or managed from the Extensions pane. There is no need
for it to be present in the preferences window as well, and indeed it is
gone from newer MacOS designs.