New preferences window
Using the preferences window from the Handy library for a more modern look
and easy implementation of a design based on the MacOS app.

The custom font setting is changed somewhat. The custom font button in the
preferences window now allows you to select a font size, which is applied
to widgets that have the .font-size-setting class. The old font size
setting is deprecated, and is replaced by the documentation font size
setting. This controls the scale of all webviews.