settings: New settings panel
Since we have already gained a dependency on the Handy library for the new
preferences window, use a preference page widget to implement the Settings
panel as well.

Preference pages automatically scroll, so this also solves the problem of
the settings panel being the bottleneck for resizing the window to a
smaller height.

Refactors it out of panel.ui into a separate widget in project-settings.ui
so that all of the logic can be put in a single file.

Also refactors how the I7App stores the retrospective data. For combo rows
it makes more sense to expose a GListStore with the data stored in GObject
instances.

Closes I7-2200 (https://inform7.atlassian.net/browse/I7-2200)