preferences: Fix problem with signal handlers outliving prefs window
Since these signal handlers were connected to a GSettings instance which
is tied to the lifetime of the app, the signal handlers were outliving the
preferences window, so attempting to change these settings after closing
and reopening the preferences window would crash.

Use g_signal_connect_object() to tie them to the lifetime of the
preferences window instead.