Introduce the Blob UI element
First introduced to the MacOS app in 2014, this is a combination status
message / progress bar that shows compilation status and lives in the
title bar.

It replaces several uses of the progress bar and status bar, which will
get removed eventually. In particular, this gets rid of the last callers
of i7_document_display_progress_busy(),
i7_document_display_status_message(), and
i7_document_remove_status_message(), so remove those.