story: Rewrite saving to be asynchronous
Saving a story involves several I/O operations, most of which don't need
to be in sequence. This changes the process to be asynchronous instead of
one-by-one blocking. See the explanation of the asynchronous operations
tree in the comment.

Now that it doesn't block the UI, we don't need to display a status
message about it.