Disable the quicksearch buttons when phrase is not found
We don't need to be able to click "Find Previous" or "Find Next" when the
phrase in the findbar entry is not present in the document. This, plus the
already existing red highlighting, should make it clear enough in the
findbar UI that the text isn't found, so we don't need to flash a message
about it additionally.