Convert the find and replace status notifications into toasts
The notifications that flashed in the main window's status bar were almost
unnoticeable since they occurred in an entirely different window than
where the user was interacting. It's probably not a good pattern that we
are flashing these notifications, and the search UI could use a good
review for best practices, but for the time being use toasts in the
find/replace window.

The toast widget is a makeshift widget using a GtkOverlay, GtkRevealer,
and GtkLabel. In GTK 4 we will be able to use AdwToast for this if we
still need it by then.