Convert the "Source code reloaded" notification into a toast
The only remaining time a transient status message was flashed, was when
the source code file was changed outside of Inform. We quietly reload it
in that case and inform the user, which is a great use case for a toast.

This is the last caller of i7_document_flash_status_message(), so we can
remove that now.