Display multiple extension download progress in the Blob
We can get rid of the old multi-download dialog, which basically consisted
only of a progress bar and a cancel button.

For this, we have to add a cancel button to the progress bar in the Blob
UI. You pass a GCancellable when you set the progress, and if it's null
the cancel button isn't available.

It's a bit iffy that we only show feedback to the user if the document
window is a story window. I guess the download code ought to be in methods
of I7Story. However, in the future with Extension Projects, those will
have a Public Library and a Blob as well, so let's keep it here for now.
Currently it's impossible to download extensions from an extension window,
so it shouldn't be a problem.