Rewrite extension downloading to be asynchronous
Downloading extension(s) didn't block the UI as such, because we iterated
the main loop in the progress callbacks, but it did make the UI slow to
respond. This changes the operations (i7_app_download_extension(),
i7_document_download_single_extension(), and
i7_document_download_multiple_extensions()) to be asynchronous.
(Downloading multiple extensions doesn't currently need a finish function,
since the UI stuff is all handled in that chain of operations.)

Installing the extension remains synchronous; that's a conversion for
another day.