search: Integrate search-and-replace dialog into search bar
We can take an example from more modern UI patterns in other GNOME apps,
and integrate the replace function into the search bar instead of having a
separate dialog for it.

With more than one row of controls, the GtkSearchBar's default close
button ends up looking weird, so first of all we add our own.

One odd thing that other GNOME apps don't usually have to deal with is
that the search bar searches the most recently focused view. This
behaviour obviously doesn't apply to searching a read-only view such as a
documentation page. If we open the search bar in replace mode, we focus
the source pane first.