story: Indicate which view is being searched
Depending on which panels are showing and which widget was last focused,
the search bar could search different views. This has the potential for
confusion. Indicate which view is being searched with a small label.

Hide the search bar's replace mode if the view being searched is not
editable.

Refactor the code for choosing which view to search, so that the panel is
responsible for picking it.