search: Refactor to reduce tight coupling
The document and search bar now are not coupled with each other. The API
of the search bar is reduced to one function: activate() which takes a
GtkWidget indicating the view being searched. The search bar performs UI
actions on the searched view.

The search bar no longer needs to keep a reference to the document, and
the document no longer needs to keep a reference to the searched view.

find_text() is no longer a vfunc of I7Document, instead there is an
activate_search() vfunc which lets the subclasses decide what to do when
the search is activated.

The search bar emits the signal 'maybe-show-entire-source' when it finds a
match which may be invisible (it can't tell by itself whether this is the
case, which is odd? But anyway) The signal handler in I7Document shows the
entire source if the match was in an invisible section.