Combine results window and find window into composite template
Previously the "find in files" dialog would open a separate dialog with
its search results. The docs search in the window's title bar would also
open this latter dialog.

The two dialogs can just as easily be combined into one. The docs search
now works like in the MacOS app: it pre-fills a search string into the
dialog and runs the search.