panel: Don't color story input text
Instead, make it bold to set it off from the rest of the text.

I was not able to find a named color in the default GTK theme that could
credibly be used as a "text accent color" legible in both light and dark
mode. However, one exists in GTK 4, so make a note to use it when we
update.

Closes: https://inform7.atlassian.net/browse/I7-2258