Move to updated Flatpak runtime
webkit2gtk-4.1 isn't in the GNOME 42 runtime.

In the GNOME 45 runtime, we need a newer version of gst-plugins-bad to
match the one in the Freedesktop 23.08 environment. Building a Flatpak on
GitHub gives some kind of git-lfs error when cloning submodules, so we
also need to skip checking out submodules; it seems like they are not
needed.

We also need to compile Inform without optimization, unfortunately,
because the version of GCC in this runtime either miscompiles something,
or there is UB in Inform that it handles differently from Clang.