Remove precompiled regexes from I7App
These are not in hot paths so they hardly need to be precompiled; and
storing them in I7App increases tight coupling, making components harder
to test individually.

Also, the plural of regex being "regices" was maybe funny to me in 2006
but now probably just impedes other people from reading the code.