Monitor game output for run-time problem messages
This is something we should have been doing all along. If the story
outputs a run-time problem message during play, we should display the
appropriate RTP page opposite the Story pane.

The way this works is going to change in the next version of Inform, but
we still need to support the old way anyway, for the retrospective Inform
versions.