Refactor I7Document file to be a construct-only property
In order to load things from the materials folder at construct time, we
need to know the location of the project on disk. This requires
refactoring the GFile pointer in I7Document so that it's set as early as
possible.

Changes the open() functions so that they use I7Document.file instead of
passing in a file.

Removes i7_document_set_file() so that the app can't get out of sync with
the location of the file on disk. The file location can only be changed
now using i7_document_save_as() with a new file.

Changes the i7_story_open() function in particular so that it doesn't try
to rename the file if it doesn't have a .inform extension. This was
probably overkill in the first place. The app doesn't make it easy to save
a file with the wrong extension, and if you manage that anyway, you can
just re-save the file yourself.