Fix double dialogs in extension download failure case
i7_app_install_extension() already displays an error dialog to the user.
Add a boolean return value to indicate whether the extension was installed
successfully or not, so that we don't get an extra success notification
after an error dialog or after the user declined permission to overwrite
an existing extension.